crewai-spitch-agent/requirements.txt
crewai 
python-dotenv
spitch
# to install this crewai tools, do pip install `crewai[tools]`
tavily-python